add table CBDT and CBLC
- These two tables are almost identical to EBDT and EBLC, so maybe we need further simplify the code
- Add NotoColorEmoji.ttf as test font

+++ tests/get-fonts.sh
@@ -15,6 +15,7 @@ for url in \
     $gh_url/adobe-fonts/source-serif-pro/release/WOFF/OTF/SourceSerifPro-Bold.otf.woff                   \
     $gh_url/adobe-fonts/source-serif-pro/release/WOFF/VAR/SourceSerifVariable-Italic.otf.woff            \
     $gh_url/alif-type/xits/master/XITSMath-Regular.otf                                                   \
+    $gh_url/googlefonts/noto-emoji/master/fonts/NotoColorEmoji.ttf                                       \
     $gh_url/googlefonts/noto-fonts/master/hinted/ttf/NotoSans/NotoSans-SemiCondensed.ttf                 \
     $gh_url/weiweihuanghuang/Work-Sans/master/fonts/static/TTF/WorkSans-Regular.ttf                      \
     https://github.com/adobe-fonts/source-han-super-otc/releases/download/20190603/SourceHanNotoCJK.ttc  \
@@ -23,7 +24,9 @@ for url in \
 
 do
     font=$(basename $url)
-    echo "Downloading $font..."
-    curl -L -o $path/$font $url
-    echo ""
+    if [ "$1" == "-f" ] || [ ! -f $path/$font ]; then
+        echo "Downloading $font..."
+        curl -L -o $path/$font $url
+        echo ""
+    fi
 done
